Moving components in the library project
