+++ projects/mifosx-app/src/app/login/login.module.ts
@@ -11,6 +11,7 @@ import { LoginFormComponent } from './login-form/login-form.component';
 import { ResetPasswordComponent } from './reset-password/reset-password.component';
 import { TwoFactorAuthenticationComponent } from './two-factor-authentication/two-factor-authentication.component';
 import { TranslateModule } from '@ngx-translate/core';
+import { CoreModule } from '@mifosx-lib/core/core.module';
 
 /**
  * Login Module
@@ -21,13 +22,17 @@ import { TranslateModule } from '@ngx-translate/core';
   imports: [
     MifosxLibModule,
     LoginRoutingModule,
-    TranslateModule
+    TranslateModule,
+    CoreModule
   ],
   declarations: [
     LoginComponent,
     LoginFormComponent,
     ResetPasswordComponent,
     TwoFactorAuthenticationComponent
+  ],
+  exports: [
+    CoreModule
   ]
 })
 export class LoginModule { }
