+++ projects/mifosx-lib/src/lib/core/authentication/authentication.service.ts
@@ -1,6 +1,6 @@
 /** Angular Imports */
 import { Injectable } from '@angular/core';
-import { HttpParams } from '@angular/common/http';
+import { HttpClient, HttpParams } from '@angular/common/http';
 
 /** rxjs Imports */
 import { Observable, of } from 'rxjs';
@@ -59,7 +59,7 @@ export class AuthenticationService {
    * @param {AlertService} alertService Alert Service.
    * @param {AuthenticationInterceptor} authenticationInterceptor Authentication Interceptor.
    */
-  constructor(private http: HttpService,
+  constructor(private http: HttpClient,
               private alertService: AlertService,
               private authenticationInterceptor: AuthenticationInterceptor) {
     this.userLoggedIn = false;
