+++ projects/mifosx-lib/src/lib/core/shell/footer/footer.component.html
