+++ projects/mifosx-lib/src/lib/core/shell/footer/footer.component.scss
@@ -1,4 +1,4 @@
-@import "../../assets/styles/helper";
+@import "../../../../assets/styles/helper";
 
 #footer {
   align-items: center;
