+++ projects/mifosx-lib/src/lib/core/shell/search-tool/search-tool.component.scss
