+++ projects/mifosx-lib/src/lib/core/shell/theme-toggle/theme-toggle.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnChanges, OnInit, SimpleChanges } from '@angular/core';
 import { ThemingService } from './theming.service';
-import { SettingsService } from '../settings/settings.service';
+import { SettingsService } from '@mifosx-lib/settings/settings.service';
 
 @Component({
   selector: 'mifosx-theme-toggle',
