--- projects/mifosx-lib/src/lib/mifosx-lib.component.ts
@@ -1,20 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'lib-mifosx-lib',
-  template: `
-    <p>
-      mifosx-lib works!
-    </p>
-  `,
-  styles: [
-  ]
-})
-export class MifosxLibComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
